requirements.txt
Django==5.2.3
djangorestframework==3.15.0
openai==1.12.0
psycopg2-binary==2.9.10 